Make CLI lazy load
This will allow parameters to be added when processing the `start` step
(for the upcomming config change).
